[Decomp][Misc] Cohesion split: extract social + instance-difficulty handlers to MiscHandler*.cpp
