+++ src/game/WorldHandlers/MiscHandler.cpp
@@ -570,254 +570,13 @@ void WorldSession::HandleStandStateChangeOpcode(WorldPacket& recv_data)
     _player->SetStandState(animstate);
 }
 
-/**
- * @brief Sends the player's friend list.
- *
- * @param recv_data The received opcode packet.
- */
-void WorldSession::HandleContactListOpcode(WorldPacket& recv_data)
-{
-    DEBUG_LOG("WORLD: Received opcode CMSG_CONTACT_LIST");
-    uint32 unk;
-    recv_data >> unk;
-    DEBUG_LOG("unk value is %u", unk);
-    _player->GetSocial()->SendSocialList();
-}
 
-/**
- * @brief Starts an asynchronous add-friend lookup.
- *
- * @param recv_data The received opcode packet.
- */
-void WorldSession::HandleAddFriendOpcode(WorldPacket& recv_data)
-{
-    DEBUG_LOG("WORLD: Received opcode CMSG_ADD_FRIEND");
 
-    std::string friendName = GetMangosString(LANG_FRIEND_IGNORE_UNKNOWN);
-    std::string friendNote;
 
-    recv_data >> friendName;
 
-    recv_data >> friendNote;
 
-    if (!normalizePlayerName(friendName))
-    {
-        return;
-    }
 
-    CharacterDatabase.escape_string(friendName);            // prevent SQL injection - normal name don't must changed by this call
 
-    DEBUG_LOG("WORLD: %s asked to add friend : '%s'",
-              GetPlayer()->GetName(), friendName.c_str());
-
-    CharacterDatabase.AsyncPQuery(&WorldSession::HandleAddFriendOpcodeCallBack, GetAccountId(), friendNote, "SELECT `guid`, `race` FROM `characters` WHERE `name` = '%s'", friendName.c_str());
-}
-
-/**
- * @brief Completes an add-friend request after the character lookup.
- *
- * @param result The async query result.
- * @param accountId The requesting account id.
- */
-void WorldSession::HandleAddFriendOpcodeCallBack(QueryResult* result, uint32 accountId, std::string friendNote)
-{
-    if (!result)
-    {
-        return;
-    }
-
-    uint32 friendLowGuid = (*result)[0].GetUInt32();
-    ObjectGuid friendGuid = ObjectGuid(HIGHGUID_PLAYER, friendLowGuid);
-    Team team = Player::TeamForRace((*result)[1].GetUInt8());
-
-    delete result;
-
-    WorldSession* session = sWorld.FindSession(accountId);
-    if (!session)
-    {
-        return;
-    }
-
-    Player* player = session->GetPlayer();
-    if (!player)
-    {
-        return;
-    }
-
-    FriendsResult friendResult = FRIEND_NOT_FOUND;
-    if (friendGuid)
-    {
-        if (friendGuid == player->GetObjectGuid())
-        {
-            friendResult = FRIEND_SELF;
-        }
-        else if (player->GetTeam() != team && !sWorld.getConfig(CONFIG_BOOL_ALLOW_TWO_SIDE_ADD_FRIEND) && session->GetSecurity() < SEC_MODERATOR)
-        {
-            friendResult = FRIEND_ENEMY;
-        }
-        else if (player->GetSocial()->HasFriend(friendGuid))
-        {
-            friendResult = FRIEND_ALREADY;
-        }
-        else
-        {
-            Player* pFriend = sObjectAccessor.FindPlayer(friendGuid);
-            if (pFriend && pFriend->IsInWorld() && pFriend->IsVisibleGloballyFor(player))
-            {
-                friendResult = FRIEND_ADDED_ONLINE;
-            }
-            else
-            {
-                friendResult = FRIEND_ADDED_OFFLINE;
-            }
-
-            if (!player->GetSocial()->AddToSocialList(friendGuid, false))
-            {
-                friendResult = FRIEND_LIST_FULL;
-                DEBUG_LOG("WORLD: %s's friend list is full.", player->GetName());
-            }
-
-            player->GetSocial()->SetFriendNote(friendGuid, friendNote);
-        }
-    }
-
-    sSocialMgr.SendFriendStatus(player, friendResult, friendGuid, false);
-
-    DEBUG_LOG("WORLD: Sent (SMSG_FRIEND_STATUS)");
-}
-
-/**
- * @brief Removes a friend from the player's social list.
- *
- * @param recv_data The received opcode packet.
- */
-void WorldSession::HandleDelFriendOpcode(WorldPacket& recv_data)
-{
-    ObjectGuid friendGuid;
-
-    DEBUG_LOG("WORLD: Received opcode CMSG_DEL_FRIEND");
-
-    recv_data >> friendGuid;
-
-    _player->GetSocial()->RemoveFromSocialList(friendGuid, false);
-
-    sSocialMgr.SendFriendStatus(GetPlayer(), FRIEND_REMOVED, friendGuid, false);
-
-    DEBUG_LOG("WORLD: Sent motd (SMSG_FRIEND_STATUS)");
-}
-
-/**
- * @brief Starts an asynchronous add-ignore lookup.
- *
- * @param recv_data The received opcode packet.
- */
-void WorldSession::HandleAddIgnoreOpcode(WorldPacket& recv_data)
-{
-    DEBUG_LOG("WORLD: Received opcode CMSG_ADD_IGNORE");
-
-    std::string IgnoreName = GetMangosString(LANG_FRIEND_IGNORE_UNKNOWN);
-
-    recv_data >> IgnoreName;
-
-    if (!normalizePlayerName(IgnoreName))
-    {
-        return;
-    }
-
-    CharacterDatabase.escape_string(IgnoreName);            // prevent SQL injection - normal name don't must changed by this call
-
-    DEBUG_LOG("WORLD: %s asked to Ignore: '%s'",
-              GetPlayer()->GetName(), IgnoreName.c_str());
-
-    CharacterDatabase.AsyncPQuery(&WorldSession::HandleAddIgnoreOpcodeCallBack, GetAccountId(), "SELECT `guid` FROM `characters` WHERE `name` = '%s'", IgnoreName.c_str());
-}
-
-/**
- * @brief Completes an add-ignore request after the character lookup.
- *
- * @param result The async query result.
- * @param accountId The requesting account id.
- */
-void WorldSession::HandleAddIgnoreOpcodeCallBack(QueryResult* result, uint32 accountId)
-{
-    if (!result)
-    {
-        return;
-    }
-
-    uint32 ignoreLowGuid = (*result)[0].GetUInt32();
-    ObjectGuid ignoreGuid = ObjectGuid(HIGHGUID_PLAYER, ignoreLowGuid);
-
-    delete result;
-
-    WorldSession* session = sWorld.FindSession(accountId);
-    if (!session)
-    {
-        return;
-    }
-
-    Player* player = session->GetPlayer();
-    if (!player)
-    {
-        return;
-    }
-
-    FriendsResult ignoreResult = FRIEND_IGNORE_NOT_FOUND;
-    if (ignoreGuid)
-    {
-        if (ignoreGuid == player->GetObjectGuid())
-        {
-            ignoreResult = FRIEND_IGNORE_SELF;
-        }
-        else if (player->GetSocial()->HasIgnore(ignoreGuid))
-        {
-            ignoreResult = FRIEND_IGNORE_ALREADY;
-        }
-        else
-        {
-            ignoreResult = FRIEND_IGNORE_ADDED;
-
-            // ignore list full
-            if (!player->GetSocial()->AddToSocialList(ignoreGuid, true))
-            {
-                ignoreResult = FRIEND_IGNORE_FULL;
-            }
-        }
-    }
-
-    sSocialMgr.SendFriendStatus(player, ignoreResult, ignoreGuid, false);
-
-    DEBUG_LOG("WORLD: Sent (SMSG_FRIEND_STATUS)");
-}
-
-/**
- * @brief Removes an ignored player from the social list.
- *
- * @param recv_data The received opcode packet.
- */
-void WorldSession::HandleDelIgnoreOpcode(WorldPacket& recv_data)
-{
-    ObjectGuid ignoreGuid;
-
-    DEBUG_LOG("WORLD: Received opcode CMSG_DEL_IGNORE");
-
-    recv_data >> ignoreGuid;
-
-    _player->GetSocial()->RemoveFromSocialList(ignoreGuid, true);
-
-    sSocialMgr.SendFriendStatus(GetPlayer(), FRIEND_IGNORE_REMOVED, ignoreGuid, false);
-
-    DEBUG_LOG("WORLD: Sent motd (SMSG_FRIEND_STATUS)");
-}
-
-void WorldSession::HandleSetContactNotesOpcode(WorldPacket& recv_data)
-{
-    DEBUG_LOG("WORLD: Received opcode CMSG_SET_CONTACT_NOTES");
-    ObjectGuid guid;
-    std::string note;
-    recv_data >> guid >> note;
-    _player->GetSocial()->SetFriendNote(guid, note);
-}
 
 /**
  * @brief Stores a bug report or suggestion from the client.
@@ -1927,125 +1686,8 @@ void WorldSession::HandleTimeSyncResp(WorldPacket& recv_data)
     _player->m_timeSyncClient = clientTicks;
 }
 
-/**
- * @brief Resets the player's or group's saved instances.
- *
- * @param recv_data The received opcode packet.
- */
-void WorldSession::HandleResetInstancesOpcode(WorldPacket& /*recv_data*/)
-{
-    DEBUG_LOG("WORLD: Received opcode CMSG_RESET_INSTANCES");
 
-    if (Group* pGroup = _player->GetGroup())
-    {
-        if (pGroup->IsLeader(_player->GetObjectGuid()))
-        {
-            pGroup->ResetInstances(INSTANCE_RESET_ALL, false, _player);
-        }
-    }
-    else
-    {
-        _player->ResetInstances(INSTANCE_RESET_ALL, false);
-    }
-}
 
-void WorldSession::HandleSetDungeonDifficultyOpcode(WorldPacket& recv_data)
-{
-    DEBUG_LOG("WORLD: Received opcode MSG_SET_DUNGEON_DIFFICULTY");
-
-    uint32 mode;
-    recv_data >> mode;
-
-    if (mode >= MAX_DUNGEON_DIFFICULTY)
-    {
-        sLog.outError("WorldSession::HandleSetDungeonDifficultyOpcode: player %d sent an invalid instance mode %d!", _player->GetGUIDLow(), mode);
-        return;
-    }
-
-    if (Difficulty(mode) == _player->GetDungeonDifficulty())
-    {
-        return;
-    }
-
-    // cannot reset while in an instance
-    Map* map = _player->GetMap();
-    if (map && map->IsDungeon())
-    {
-        sLog.outError("WorldSession::HandleSetDungeonDifficultyOpcode: player %d tried to reset the instance while inside!", _player->GetGUIDLow());
-        return;
-    }
-
-    // Exception to set mode to normal for low-level players
-    if (_player->getLevel() < LEVELREQUIREMENT_HEROIC && mode > REGULAR_DIFFICULTY)
-    {
-        return;
-    }
-
-    if (Group* pGroup = _player->GetGroup())
-    {
-        if (pGroup->IsLeader(_player->GetObjectGuid()))
-        {
-            // the difficulty is set even if the instances can't be reset
-            //_player->SendDungeonDifficulty(true);
-            pGroup->ResetInstances(INSTANCE_RESET_CHANGE_DIFFICULTY, false, _player);
-            pGroup->SetDungeonDifficulty(Difficulty(mode));
-        }
-    }
-    else
-    {
-        _player->ResetInstances(INSTANCE_RESET_CHANGE_DIFFICULTY, false);
-        _player->SetDungeonDifficulty(Difficulty(mode));
-    }
-}
-
-void WorldSession::HandleSetRaidDifficultyOpcode(WorldPacket& recv_data)
-{
-    DEBUG_LOG("WORLD: Received opcode MSG_SET_RAID_DIFFICULTY");
-
-    uint32 mode;
-    recv_data >> mode;
-
-    if (mode >= MAX_RAID_DIFFICULTY)
-    {
-        sLog.outError("WorldSession::HandleSetRaidDifficultyOpcode: player %d sent an invalid instance mode %d!", _player->GetGUIDLow(), mode);
-        return;
-    }
-
-    if (Difficulty(mode) == _player->GetRaidDifficulty())
-    {
-        return;
-    }
-
-    // cannot reset while in an instance
-    Map* map = _player->GetMap();
-    if (map && map->IsDungeon())
-    {
-        sLog.outError("WorldSession::HandleSetRaidDifficultyOpcode: player %d tried to reset the instance while inside!", _player->GetGUIDLow());
-        return;
-    }
-
-    // Exception to set mode to normal for low-level players
-    if (_player->getLevel() < LEVELREQUIREMENT_HEROIC && mode > REGULAR_DIFFICULTY)
-    {
-        return;
-    }
-
-    if (Group* pGroup = _player->GetGroup())
-    {
-        if (pGroup->IsLeader(_player->GetObjectGuid()))
-        {
-            // the difficulty is set even if the instances can't be reset
-            //_player->SendDungeonDifficulty(true);
-            pGroup->ResetInstances(INSTANCE_RESET_CHANGE_DIFFICULTY, true, _player);
-            pGroup->SetRaidDifficulty(Difficulty(mode));
-        }
-    }
-    else
-    {
-        _player->ResetInstances(INSTANCE_RESET_CHANGE_DIFFICULTY, true);
-        _player->SetRaidDifficulty(Difficulty(mode));
-    }
-}
 
 /**
  * @brief Cancels the player's mount aura when allowed.
